# Conflicts:
#	vlib/v/parser/checks.v

+++ examples/term.ui/pong.v
@@ -212,7 +212,7 @@ fn (mut p Player) update() {
 	// dt := p.game.app.dt
 	ball := unsafe { &p.game.ball }
 	// Evil AI that eventually will take over the world
-	p.pos.y = ball.pos.y - int(f32(p.racket_size) * 0.5)
+	p.pos.y = ball.pos.y - f32(p.racket_size) * 0.5
 }
 
 struct Vec {
@@ -278,7 +278,7 @@ fn (mut g Game) reset() {
 fn (mut g Game) new_round() {
 	mut i := 0
 	for mut p in g.players {
-		p.pos.x = if i == 0 { 3 } else { g.app.width - 2 }
+		p.pos.x = if i == 0 { f32(3) } else { f32(g.app.width - 2) }
 		p.pos.y = f32(g.app.height) * 0.5 - f32(p.racket_size) * 0.5
 		i++
 	}
@@ -294,10 +294,10 @@ fn (mut g Game) update() {
 		p.update()
 		// Keep rackets within the game area
 		if p.pos.y <= 0 {
-			p.pos.y = 1
+			p.pos.y = 1.0
 		}
 		if p.pos.y + p.racket_size >= g.app.height {
-			p.pos.y = g.app.height - p.racket_size - 1
+			p.pos.y = f32(g.app.height - p.racket_size - 1)
 		}
 		// Check ball collision
 		// Player left side
